refactor: boot selection with name and sku
