+++ client/src/components/forms/index.js
@@ -2,6 +2,6 @@ import CustomerForm from './CustomerForm.component';
 import EventForm from './EventForm.component';
 import EventDetailsForm from './EventDetailsForm.component';
 import JoinForm from './JoinForm.component';
+import BootSelect from './bootSelect/BootSelect.component';
 
-
-export { CustomerForm, EventForm, EventDetailsForm, JoinForm };
+export { CustomerForm, EventForm, EventDetailsForm, JoinForm, BootSelect };
