+++ client/src/components/pageElements/index.js
@@ -2,7 +2,7 @@ import Header from './Header.component';
 import Footer from './Footer.component';
 import EventDisplay from './EventDisplay.component';
 import GuestsDisplay from './GuestsDisplay.component';
-import BootCarousel from './BootCarousel.component';
+import BootSelect from '../forms/bootSelect/BootSelect.component';
 
 
-export { Header, Footer, EventDisplay, GuestsDisplay, BootCarousel };
+export { Header, Footer, EventDisplay, GuestsDisplay, BootSelect };
